simplicity haven’t hooked up the database yet. doing that now. I installed a global mysql which is conflicting with my xampp mysql
simplicity is a template library. not to be confused with a template engine. it will take premade templates but that’s not it’s primary function. it’s built to make it easy to create a template to further separate content from code. This means that people who don’t know code will be limited as to how much they can do in terms of moving content around  but sometimes that just for the best.